Phase 1: extract features/diagrams.js (mermaid render + panzoom + fullscreen engine)
Moves the diagram engine — processMermaidDiagrams, createDiagramContainer,
fitDiagramToContainer, resetToFit, updateZoomUI, initializePanzoom,
openFullscreen, setupFullscreenControls, closeFullscreen, cleanupPanzoomInstances,
reRenderMermaidDiagrams — into one module (showDropZone stays in main). It's
import-only (Panzoom/mermaid/DOMPurify + state/platform + getSvgNaturalDimensions
/getMermaidConfig + minimap/diagram-export/share-links); fullscreenOverlay and
markdownContent are looked up by id. main exports-in: the 6 functions it still
calls (processMermaidDiagrams, cleanupPanzoomInstances, reRenderMermaidDiagrams,
closeFullscreen, updateZoomUI, resetToFit).

main.js ~1,544 -> ~1,017 lines (from 2,814 at the split's start).
Build + lint clean; 297 tests pass.
